pyproject.toml
[tool.poetry]
name = "streaming_playground"
version = "0.1.0"
description = ""
authors = ["Maher Harring Kassem <maher.kassem@maersk.com>"]

[tool.poetry.dependencies]
python = "^3.9"
delta-spark = "^2.2"
pyspark = "3.4.1"
confluent-kafka = "^2.1.1"
pytest = "^7.4.0"
black = "^23.3.0"
isort = "^5.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
